feat: extract entries download logic to a new service
